# openapi.json
{"openapi": "3.1.0", "info": {"title": "Frill API", "version": "0.1.0"}, "paths": {"/auth/signup": {"post": {"tags": ["auth"], "summary": "Signup", "operationId": "signup_auth_signup_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/SignupRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "operationId": "login_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/me": {"get": {"tags": ["auth"], "summary": "Me", "operationId": "me_auth_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}}}}, "/auth/logout": {"post": {"tags": ["auth"], "summary": "Logout", "operationId": "logout_auth_logout_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/auth/profile": {"put": {"tags": ["auth"], "summary": "Update Profile", "operationId": "update_profile_auth_profile_put", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProfileUpdateRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/forgot-password": {"post": {"tags": ["auth"], "summary": "Forgot Password", "operationId": "forgot_password_auth_forgot_password_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ForgotPasswordRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/reset-password": {"post": {"tags": ["auth"], "summary": "Reset Password", "operationId": "reset_password_auth_reset_password_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/ResetPasswordRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/products": {"get": {"tags": ["products"], "summary": "List Products", "operationId": "list_products_api_products_get", "parameters": [{"name": "category", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Category"}}, {"name": "sort", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Sort"}}, {"name": "q", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Q"}}, {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "default": 1, "title": "Page"}}, {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "default": 20, "title": "Page Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedProducts"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["products"], "summary": "Create Product", "operationId": "create_product_api_products_post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/products/{slug}": {"get": {"tags": ["products"], "summary": "Get Product By Slug", "operationId": "get_product_by_slug_api_products__slug__get", "parameters": [{"name": "slug", "in": "path", "required": true, "schema": {"type": "string", "title": "Slug"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/products/id/{product_id}": {"get": {"tags": ["products"], "summary": "Get Product By Id", "operationId": "get_product_by_id_api_products_id__product_id__get", "parameters": [{"name": "product_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Product Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/products/{product_id}": {"patch": {"tags": ["products"], "summary": "Update Product", "operationId": "update_product_api_products__product_id__patch", "parameters": [{"name": "product_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Product Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["products"], "summary": "Delete Product", "operationId": "delete_product_api_products__product_id__delete", "parameters": [{"name": "product_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Product Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/designs": {"post": {"tags": ["designs"], "summary": "Save Design", "operationId": "save_design_api_designs_post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DesignCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DesignOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["designs"], "summary": "List Designs", "operationId": "list_designs_api_designs_get", "parameters": [{"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "default": 1, "title": "Page"}}, {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "default": 20, "title": "Page Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedDesigns"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/designs/{design_id}": {"delete": {"tags": ["designs"], "summary": "Delete Design", "operationId": "delete_design_api_designs__design_id__delete", "parameters": [{"name": "design_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Design Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/designs/{design_id}/status": {"patch": {"tags": ["designs"], "summary": "Update Design Status", "operationId": "update_design_status_api_designs__design_id__status_patch", "parameters": [{"name": "design_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Design Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DesignStatusUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DesignOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/orders": {"post": {"tags": ["orders"], "summary": "Create Order", "operationId": "create_order_api_orders_post", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateOrderRequest"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateOrderResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "get": {"tags": ["orders"], "summary": "List Orders", "operationId": "list_orders_api_orders_get", "parameters": [{"name": "status", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Status"}}, {"name": "page", "in": "query", "required": false, "schema": {"type": "integer", "minimum": 1, "default": 1, "title": "Page"}}, {"name": "page_size", "in": "query", "required": false, "schema": {"type": "integer", "maximum": 100, "minimum": 1, "default": 20, "title": "Page Size"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PaginatedOrders"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/orders/stats": {"get": {"tags": ["orders"], "summary": "Admin Stats", "description": "FIX C2 \u2014 Admin stats endpoint (was missing entirely).\nReplaces the hardcoded KPI data in AdminDashboardPage.jsx.", "operationId": "admin_stats_api_orders_stats_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AdminStatsResponse"}}}}}}}, "/api/orders/{order_id}": {"get": {"tags": ["orders"], "summary": "Get Order", "operationId": "get_order_api_orders__order_id__get", "parameters": [{"name": "order_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Order Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OrderOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/orders/{order_id}/status": {"patch": {"tags": ["orders"], "summary": "Update Order Status", "operationId": "update_order_status_api_orders__order_id__status_patch", "parameters": [{"name": "order_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid", "title": "Order Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateOrderStatus"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OrderOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/upload/image": {"post": {"tags": ["upload"], "summary": "Upload Image", "operationId": "upload_image_api_upload_image_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_upload_image_api_upload_image_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/upload/design-export": {"post": {"tags": ["upload"], "summary": "Upload Design Export", "operationId": "upload_design_export_api_upload_design_export_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_upload_design_export_api_upload_design_export_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/upload/product-image": {"post": {"tags": ["upload"], "summary": "Upload Product Image", "operationId": "upload_product_image_api_upload_product_image_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_upload_product_image_api_upload_product_image_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/health": {"get": {"summary": "Health", "operationId": "health_health_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"AdminStatsResponse": {"properties": {"total_revenue": {"type": "number", "title": "Total Revenue"}, "total_orders": {"type": "integer", "title": "Total Orders"}, "orders_today": {"type": "integer", "title": "Orders Today"}, "active_products": {"type": "integer", "title": "Active Products"}, "total_customers": {"type": "integer", "title": "Total Customers"}}, "type": "object", "required": ["total_revenue", "total_orders", "orders_today", "active_products", "total_customers"], "title": "AdminStatsResponse"}, "Body_upload_design_export_api_upload_design_export_post": {"properties": {"mockup": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Mockup"}, "print_file": {"type": "string", "contentMediaType": "application/octet-stream", "title": "Print File"}}, "type": "object", "required": ["mockup", "print_file"], "title": "Body_upload_design_export_api_upload_design_export_post"}, "Body_upload_image_api_upload_image_post": {"properties": {"file": {"type": "string", "contentMediaType": "application/octet-stream", "title": "File"}}, "type": "object", "required": ["file"], "title": "Body_upload_image_api_upload_image_post"}, "Body_upload_product_image_api_upload_product_image_post": {"properties": {"file": {"type": "string", "contentMediaType": "application/octet-stream", "title": "File"}}, "type": "object", "required": ["file"], "title": "Body_upload_product_image_api_upload_product_image_post"}, "CreateOrderRequest": {"properties": {"first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "email": {"type": "string", "format": "email", "title": "Email"}, "phone": {"type": "string", "title": "Phone"}, "address": {"type": "string", "title": "Address"}, "city": {"type": "string", "title": "City"}, "province": {"type": "string", "title": "Province"}, "postal_code": {"type": "string", "title": "Postal Code"}, "payment_method": {"type": "string", "enum": ["cod", "jazzcash", "easypaisa", "stripe"], "title": "Payment Method"}, "wallet_number": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Wallet Number"}, "items": {"items": {"$ref": "#/components/schemas/OrderItemIn"}, "type": "array", "title": "Items"}}, "type": "object", "required": ["first_name", "last_name", "email", "phone", "address", "city", "province", "postal_code", "payment_method", "items"], "title": "CreateOrderRequest"}, "CreateOrderResponse": {"properties": {"order": {"$ref": "#/components/schemas/OrderOut"}, "payment_metadata": {"additionalProperties": true, "type": "object", "title": "Payment Metadata"}}, "type": "object", "required": ["order", "payment_metadata"], "title": "CreateOrderResponse"}, "DesignCreate": {"properties": {"name": {"type": "string", "title": "Name"}, "design_json": {"additionalProperties": true, "type": "object", "title": "Design Json"}, "product_id": {"type": "string", "title": "Product Id"}, "color_id": {"type": "string", "title": "Color Id"}, "view_id": {"type": "string", "title": "View Id"}, "mockup_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Mockup Url"}}, "type": "object", "required": ["name", "design_json", "product_id", "color_id", "view_id"], "title": "DesignCreate"}, "DesignOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "user_id": {"type": "string", "format": "uuid", "title": "User Id"}, "name": {"type": "string", "title": "Name"}, "design_json": {"additionalProperties": true, "type": "object", "title": "Design Json"}, "product_id": {"type": "string", "format": "uuid", "title": "Product Id"}, "color_id": {"type": "string", "title": "Color Id"}, "view_id": {"type": "string", "title": "View Id"}, "mockup_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Mockup Url"}, "status": {"type": "string", "title": "Status"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["id", "user_id", "name", "design_json", "product_id", "color_id", "view_id", "mockup_url", "status", "created_at", "updated_at"], "title": "DesignOut"}, "DesignStatusUpdate": {"properties": {"status": {"type": "string", "title": "Status"}}, "type": "object", "required": ["status"], "title": "DesignStatusUpdate"}, "ForgotPasswordRequest": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}}, "type": "object", "required": ["email"], "title": "ForgotPasswordRequest"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "LoginRequest": {"properties": {"email": {"type": "string", "format": "email", "title": "Email"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["email", "password"], "title": "LoginRequest"}, "OrderItemIn": {"properties": {"product_id": {"type": "string", "title": "Product Id"}, "quantity": {"type": "integer", "title": "Quantity"}, "selected_size": {"type": "string", "title": "Selected Size"}, "selected_color_name": {"type": "string", "title": "Selected Color Name"}, "selected_color_hex": {"type": "string", "title": "Selected Color Hex"}, "selected_view_label": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Selected View Label"}, "mockup_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Mockup Url"}, "print_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Print Url"}, "design_json": {"anyOf": [{"additionalProperties": true, "type": "object"}, {"type": "null"}], "title": "Design Json"}}, "type": "object", "required": ["product_id", "quantity", "selected_size", "selected_color_name", "selected_color_hex"], "title": "OrderItemIn"}, "OrderItemOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "product_id": {"anyOf": [{"type": "string", "format": "uuid"}, {"type": "null"}], "title": "Product Id"}, "product_name_snapshot": {"type": "string", "title": "Product Name Snapshot"}, "price_snapshot": {"type": "number", "title": "Price Snapshot"}, "quantity": {"type": "integer", "title": "Quantity"}, "selected_size": {"type": "string", "title": "Selected Size"}, "selected_color_name": {"type": "string", "title": "Selected Color Name"}, "selected_color_hex": {"type": "string", "title": "Selected Color Hex"}, "selected_view_label": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Selected View Label"}, "mockup_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Mockup Url"}, "print_url": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Print Url"}}, "type": "object", "required": ["id", "product_id", "product_name_snapshot", "price_snapshot", "quantity", "selected_size", "selected_color_name", "selected_color_hex", "selected_view_label", "mockup_url", "print_url"], "title": "OrderItemOut"}, "OrderOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "user_id": {"type": "string", "format": "uuid", "title": "User Id"}, "first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "email": {"type": "string", "title": "Email"}, "phone": {"type": "string", "title": "Phone"}, "address": {"type": "string", "title": "Address"}, "city": {"type": "string", "title": "City"}, "province": {"type": "string", "title": "Province"}, "postal_code": {"type": "string", "title": "Postal Code"}, "payment_method": {"type": "string", "title": "Payment Method"}, "wallet_number": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Wallet Number"}, "status": {"type": "string", "title": "Status"}, "total": {"type": "number", "title": "Total"}, "items": {"items": {"$ref": "#/components/schemas/OrderItemOut"}, "type": "array", "title": "Items"}, "created_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Created At"}, "item_count": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Item Count"}}, "type": "object", "required": ["id", "user_id", "first_name", "last_name", "email", "phone", "address", "city", "province", "postal_code", "payment_method", "wallet_number", "status", "total", "items"], "title": "OrderOut"}, "PaginatedDesigns": {"properties": {"items": {"items": {"$ref": "#/components/schemas/DesignOut"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}, "page": {"type": "integer", "title": "Page"}, "page_size": {"type": "integer", "title": "Page Size"}}, "type": "object", "required": ["items", "total", "page", "page_size"], "title": "PaginatedDesigns"}, "PaginatedOrders": {"properties": {"items": {"items": {"$ref": "#/components/schemas/OrderOut"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}, "page": {"type": "integer", "title": "Page"}, "page_size": {"type": "integer", "title": "Page Size"}}, "type": "object", "required": ["items", "total", "page", "page_size"], "title": "PaginatedOrders"}, "PaginatedProducts": {"properties": {"items": {"items": {"$ref": "#/components/schemas/ProductOut"}, "type": "array", "title": "Items"}, "total": {"type": "integer", "title": "Total"}, "page": {"type": "integer", "title": "Page"}, "page_size": {"type": "integer", "title": "Page Size"}}, "type": "object", "required": ["items", "total", "page", "page_size"], "title": "PaginatedProducts"}, "PrintArea": {"properties": {"x": {"type": "number", "title": "X"}, "y": {"type": "number", "title": "Y"}, "width": {"type": "number", "title": "Width"}, "height": {"type": "number", "title": "Height"}}, "type": "object", "required": ["x", "y", "width", "height"], "title": "PrintArea"}, "ProductColorIn": {"properties": {"name": {"type": "string", "title": "Name"}, "hex": {"type": "string", "title": "Hex"}, "views": {"items": {"$ref": "#/components/schemas/ProductViewIn"}, "type": "array", "title": "Views", "default": []}}, "type": "object", "required": ["name", "hex"], "title": "ProductColorIn"}, "ProductColorOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "hex": {"type": "string", "title": "Hex"}, "views": {"items": {"$ref": "#/components/schemas/ProductViewOut"}, "type": "array", "title": "Views"}}, "type": "object", "required": ["id", "name", "hex", "views"], "title": "ProductColorOut"}, "ProductCreate": {"properties": {"slug": {"type": "string", "title": "Slug"}, "name": {"type": "string", "title": "Name"}, "vendor": {"type": "string", "title": "Vendor"}, "category": {"type": "string", "title": "Category"}, "description": {"type": "string", "title": "Description"}, "long_description": {"type": "string", "title": "Long Description", "default": ""}, "price": {"type": "number", "title": "Price"}, "old_price": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Old Price"}, "customizable": {"type": "boolean", "title": "Customizable", "default": false}, "sizes": {"items": {"type": "string"}, "type": "array", "title": "Sizes", "default": []}, "colors": {"items": {"$ref": "#/components/schemas/ProductColorIn"}, "type": "array", "title": "Colors", "default": []}}, "type": "object", "required": ["slug", "name", "vendor", "category", "description", "price"], "title": "ProductCreate"}, "ProductOut": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "slug": {"type": "string", "title": "Slug"}, "name": {"type": "string", "title": "Name"}, "vendor": {"type": "string", "title": "Vendor"}, "category": {"type": "string", "title": "Category"}, "description": {"type": "string", "title": "Description"}, "long_description": {"type": "string", "title": "Long Description"}, "price": {"type": "number", "title": "Price"}, "old_price": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Old Price"}, "stars": {"type": "number", "title": "Stars"}, "review_count": {"type": "integer", "title": "Review Count"}, "customizable": {"type": "boolean", "title": "Customizable"}, "sizes": {"items": {"type": "string"}, "type": "array", "title": "Sizes"}, "colors": {"items": {"$ref": "#/components/schemas/ProductColorOut"}, "type": "array", "title": "Colors"}, "created_at": {"anyOf": [{"type": "string", "format": "date-time"}, {"type": "null"}], "title": "Created At"}, "img": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Img", "readOnly": true}}, "type": "object", "required": ["id", "slug", "name", "vendor", "category", "description", "long_description", "price", "old_price", "stars", "review_count", "customizable", "sizes", "colors", "img"], "title": "ProductOut"}, "ProductUpdate": {"properties": {"name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}, "vendor": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Vendor"}, "category": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Category"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "long_description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Long Description"}, "price": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Price"}, "old_price": {"anyOf": [{"type": "number"}, {"type": "null"}], "title": "Old Price"}, "customizable": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Customizable"}, "sizes": {"anyOf": [{"items": {"type": "string"}, "type": "array"}, {"type": "null"}], "title": "Sizes"}}, "type": "object", "title": "ProductUpdate"}, "ProductViewIn": {"properties": {"label": {"type": "string", "title": "Label"}, "image_url": {"type": "string", "title": "Image Url"}, "print_area": {"$ref": "#/components/schemas/PrintArea"}}, "type": "object", "required": ["label", "image_url", "print_area"], "title": "ProductViewIn"}, "ProductViewOut": {"properties": {"label": {"type": "string", "title": "Label"}, "image_url": {"type": "string", "title": "Image Url"}, "print_area": {"$ref": "#/components/schemas/PrintArea"}, "id": {"type": "string", "format": "uuid", "title": "Id"}}, "type": "object", "required": ["label", "image_url", "print_area", "id"], "title": "ProductViewOut"}, "ProfileUpdateRequest": {"properties": {"first_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "First Name"}, "last_name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Last Name"}, "phone": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Phone"}}, "type": "object", "title": "ProfileUpdateRequest"}, "ResetPasswordRequest": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "new_password": {"type": "string", "title": "New Password"}}, "type": "object", "required": ["access_token", "new_password"], "title": "ResetPasswordRequest"}, "SignupRequest": {"properties": {"first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "email": {"type": "string", "format": "email", "title": "Email"}, "phone": {"type": "string", "title": "Phone"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["first_name", "last_name", "email", "phone", "password"], "title": "SignupRequest"}, "UpdateOrderStatus": {"properties": {"status": {"type": "string", "enum": ["pending", "processing", "shipped", "delivered", "cancelled"], "title": "Status"}}, "type": "object", "required": ["status"], "title": "UpdateOrderStatus"}, "UserResponse": {"properties": {"id": {"type": "string", "format": "uuid", "title": "Id"}, "first_name": {"type": "string", "title": "First Name"}, "last_name": {"type": "string", "title": "Last Name"}, "email": {"type": "string", "title": "Email"}, "phone": {"type": "string", "title": "Phone"}, "role": {"type": "string", "title": "Role"}}, "type": "object", "required": ["id", "first_name", "last_name", "email", "phone", "role"], "title": "UserResponse"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}